Updated tests for createStaticFacet and createFacetContext
