+++ packages/@react-facet/core/src/createFacetContext.spec.tsx
@@ -2,7 +2,7 @@ import React, { useContext } from 'react'
 import { act, render } from '@react-facet/dom-fiber-testing-library'
 import { createFacetContext } from './createFacetContext'
 import { useFacetState } from './hooks'
-import { Setter } from './types'
+import { Facet, Setter } from './types'
 
 it(`has default value`, () => {
   const defaultValue = 'defaultValue'
@@ -70,4 +70,60 @@ it(`it updates the context value without re-conciliation`, () => {
 
   expect(result.baseElement).toContainHTML(evenNewerValue)
   expect(hasRenderedMock).toBeCalledTimes(1)
+})
+
+// Tests that the intial value of the createFacetContext is static
+describe('createFacetContext initial facet', () => {
+  const env = process.env
+  let staticFacet: Facet<string>
+  const defaultValue = 'defaultValue'
+
+  beforeEach(() => {
+    jest.resetModules()
+    process.env = { ...env }
+
+    const context = createFacetContext(defaultValue)
+    const Component = () => {
+      const stringFacet = useContext(context)
+      staticFacet = stringFacet
+      return <fast-text text={stringFacet}></fast-text>
+    }
+
+    render(<Component />)
+  })
+
+  afterEach(() => {
+    process.env = env
+  })
+
+  it(`it can be read but not mutated`, () => {
+    act(() => {
+      expect(staticFacet.get()).toBe(defaultValue)
+      expect('set' in staticFacet).toBe(false)
+    })
+  })
+
+  it(`it responds with the same value if you observe it and warns you in a non-production environment`, () => {
+    const consoleLogMock = jest.spyOn(console, 'log').mockImplementation()
+
+    const update = jest.fn()
+
+    staticFacet.observe(update)
+    expect(update).toHaveBeenCalledTimes(1)
+    expect(update).toHaveBeenCalledWith(defaultValue)
+    expect(consoleLogMock).toHaveBeenCalledTimes(1)
+    expect(consoleLogMock).toHaveBeenCalledWith(
+      `Accessing a static facet created through createFacetContext, perhaps you're missing a Context Provider?`,
+    )
+
+    update.mockClear()
+    consoleLogMock.mockClear()
+
+    process.env.NODE_ENV = 'production'
+
+    staticFacet.observe(update)
+    expect(update).toHaveBeenCalledTimes(1)
+    expect(update).toHaveBeenCalledWith(defaultValue)
+    expect(consoleLogMock).toHaveBeenCalledTimes(0)
+  })
 })
